pyproject.toml
[project]
name = "simpleedgegateway"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiosqlite>=0.22.1",
    "fastapi>=0.128.0",
    "jinja2>=3.1.6",
    "pymodbus>=3.11.4",
    "python-multipart>=0.0.21",
    "pyyaml>=6.0.3",
    "uvicorn>=0.40.0",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.uv]
package = true

[project.scripts]
web-gui = "src.web.webmain:main"
poller = "src.engine.poller:main"
dbinit = "src.common.db_handler:init_db"
